+++ frontend/src/components/AddPlatform/AddPlatform.tsx
